Muevo helpers de config a módulo aparte
