+++ ckanext/gobar_theme/helpers/__init__.py
@@ -7,3 +7,4 @@ from .types import *
 from .cron import *
 from .distribution import *
 from .dates import *
+from .config import *
