+++ ckanext/gobar_theme/helpers/helpers.py
@@ -1,18 +1,14 @@
 #!coding=utf-8
-import json
 import csv
+import json
 import logging
-import os
 import subprocess
 from urlparse import urljoin
 from urlparse import urlparse
 
-from pylons import config as config
 import ckan.lib.helpers as ckan_helpers
 import ckan.logic as logic
-from ckan.common import request, _
-from ckanext import constants
-from ckanext.gobar_theme.theme_config import ThemeConfig
+from ckan.common import request
 
 logger = logging.getLogger(__name__)
 
@@ -49,11 +45,6 @@ def cut_text(text, limit):
 
 def cut_img_path(url):
     return urlparse(url).path
-
-
-def get_theme_config(path=None, default=None):
-    theme_config = ThemeConfig(constants.CONFIG_PATH)
-    return theme_config.get(path, default)
 
 
 def url_join(base, url, *args):
@@ -119,61 +110,8 @@ def jsondump(field=''):
     return Markup(json.dumps(field))
 
 
-def get_default_background_configuration():
-    background_opacity = config.get('andino.background_opacity')
-    return background_opacity
-
-
-def get_gtm_code():
-    return get_theme_config('google_tag_manager.container-id') or \
-           config.get('ckan.google_tag_manager.gtm_container_id', '')
-
-
-def get_current_url_for_resource(package_id, resource_id):
-    return os.path.join(config.get('ckan.site_url'), 'dataset', package_id, 'resource', resource_id)
-
-
-def get_package_organization(package_id):
-    return logic.get_action('package_show')({}, {'id': package_id}).get('organization', {})
-
-
-def get_resource_icon(resource):
-    icon_url = resource.get('icon_url', None)
-    if icon_url:
-        return icon_url
-    package_id = resource['package_id']
-    id_to_search_with = '%s_%s_%s' % (
-        get_package_organization(package_id).get('id', ''),
-        resource['package_id'],
-        resource['id']
-    )
-    resource_in_config = get_theme_config('resources', {}).get(id_to_search_with, None)
-    if resource_in_config is not None:
-        return resource_in_config.get('icon_url', None)
-    return None
-
-
-def get_andino_base_page():
-    return config.get('andino.base_page', 'gobar_page.html')
-
-
-def get_default_series_api_url():
-    return config.get('seriestiempoarexplorer.default_series_api_uri', '')
-
-
 def get_current_terminal_username():
     return subprocess.check_output("whoami").strip()
-
-
-def search_for_value_in_config_file(field):
-    # Solamente queremos utilizar el valor default cuando no existe uno ingresado por el usuario.
-    try:
-        value = subprocess.check_output(
-            'grep -E "^{}[[:space:]]*=[[:space:]]*" '
-            '/etc/ckan/default/production.ini | tr -d [[:space:]]'.format(field), shell=True).strip()
-        return value.replace(field, '')[1:]
-    except Exception:
-        return ''
 
 
 def delete_column_from_csv_file(csv_path, column_name):
@@ -194,8 +132,3 @@ def delete_column_from_csv_file(csv_path, column_name):
         wtr = csv.writer(result)
         for r in list_with_rows:
             wtr.writerow(tuple(x for x in r))
-
-
-def is_plugin_present(plugin_name):
-    plugins = config.get('ckan.plugins')
-    return plugin_name in plugins
